Board 로직 추가, 게시글 상세보기 구현 완료
