+++ src/main/java/com/termproject/board/dto/ResponseBoardDto.java
@@ -1,7 +1,11 @@
 package com.termproject.board.dto;
 
 import com.termproject.board.domain.board.Board;
+import com.termproject.board.domain.comment.Comment;
 import lombok.*;
+
+import java.sql.Timestamp;
+import java.util.List;
 
 @Getter
 @Setter
@@ -9,9 +13,14 @@ import lombok.*;
 @NoArgsConstructor
 @Builder
 public class ResponseBoardDto {
+
+    private Timestamp createTime;
+    private Timestamp modifiedDate;
     private String title;
     private String content;
     private int likes;
+
+    private List<Comment> comments;
 
     public ResponseBoardDto(Board board){
         setData(board);
@@ -21,5 +30,8 @@ public class ResponseBoardDto {
         this.title = board.getTitle();
         this.content = board.getContent();
         this.likes = board.getLikes();
+        this.comments = board.getComments();
+        this.createTime = board.getCreateTime();
+        this.modifiedDate = board.getModifiedDate();
     }
 }
